+++ pagurus_windows_system/src/lib.rs
@@ -3,3 +3,4 @@ mod system;
 mod window;
 
 pub use self::system::{WindowsSystem, WindowsSystemBuilder};
+pub use self::window::Window;